Restoring back to index button, test context clean up
