+++ src/components/BrowseForm/BrowseForm.test.tsx
@@ -14,6 +14,7 @@ describe("BrowseForm", () => {
     )
 
   beforeEach(() => {
+    mockRouter.push("/", undefined, { shallow: true })
     mockRouter.query = {}
   })
 
@@ -41,7 +42,7 @@ describe("BrowseForm", () => {
     const select = screen.getByLabelText("Select a category")
 
     await userEvent.type(input, "Vietnam War")
-    await userEvent.selectOptions(select, "starts_with")
+    await userEvent.selectOptions(select, "subject_starts_with")
     submit()
     expect(mockRouter.asPath).toBe(
       "/browse?q=Vietnam+War&search_scope=starts_with&sort=termLabel&sort_direction=asc"
@@ -75,12 +76,48 @@ describe("BrowseForm", () => {
     const select = screen.getByLabelText(
       "Select a category"
     ) as HTMLSelectElement
-    expect(select.value).toBe("starts_with")
+    expect(select.value).toBe("subject_starts_with")
   })
 
   it("renders the browse tip", () => {
     render(component)
     const tip = screen.getByText(/browse tip/i)
     expect(tip).toBeInTheDocument()
+  })
+
+  it("updates the browse type when changing the dropdown", async () => {
+    render(component)
+    const select = screen.getByLabelText("Select a category")
+
+    // Default is subjects
+    expect(mockRouter.asPath).toBe("/")
+
+    await userEvent.selectOptions(select, "contributor_has")
+
+    expect(mockRouter.asPath).toBe("/browse/authors?search_scope=has")
+  })
+
+  it("maintains the search term when changing browse type", async () => {
+    render(component)
+    const input = screen.getByRole("textbox")
+    const select = screen.getByLabelText("Select a category")
+
+    await userEvent.type(input, "Ornithology")
+    expect(input).toHaveValue("Ornithology")
+
+    await userEvent.selectOptions(select, "subject_starts_with")
+
+    expect(input).toHaveValue("Ornithology")
+  })
+
+  it("updates URL correctly when switching between subjects and contributors", async () => {
+    render(component)
+    const select = screen.getByLabelText("Select a category")
+
+    await userEvent.selectOptions(select, "contributor_starts_with")
+    expect(mockRouter.asPath).toBe("/browse/authors?search_scope=starts_with")
+
+    await userEvent.selectOptions(select, "subject_has")
+    expect(mockRouter.asPath).toBe("/browse?search_scope=has")
   })
 })
