--- src/components/SearchBrowseForm/SearchBrowseForm.tsx
@@ -1,190 +1,0 @@
-import {
-  Box,
-  Flex,
-  Icon,
-  Link,
-  SearchBar,
-  Text,
-} from "@nypl/design-system-react-components"
-import { useRouter } from "next/router"
-import { useEffect, useState, type SyntheticEvent } from "react"
-import { useFocusContext } from "../../context/FocusContext"
-import useLoading from "../../hooks/useLoading"
-import styles from "../../../styles/components/Search.module.scss"
-import type { RCPage } from "../../types/pageTypes"
-import type { BrowseType } from "../../types/browseTypes"
-import { useBrowseContext } from "../../context/BrowseContext"
-import { getBrowseFormKey } from "../../utils/browseUtils"
-
-type BrowseFormOption = {
-  browseType: BrowseType
-  scope: string
-  text: string
-  searchTip?: string | JSX.Element
-  placeholder: string
-}
-
-type SearchBrowseFormProps = {
-  path: string
-  tipTitle: string
-  labelText?: string
-  selectOptions: { [key: string]: BrowseFormOption }
-  getQueryString: (params: { [key: string]: string }) => string
-  onSubmitFocusId?: string | null
-  children?: React.ReactNode
-  activePage: RCPage
-}
-
-const SearchBrowseForm = ({
-  path,
-  tipTitle,
-  labelText = "Search Bar Label",
-  selectOptions,
-  getQueryString,
-  onSubmitFocusId = null,
-  children,
-  activePage,
-}: SearchBrowseFormProps) => {
-  const router = useRouter()
-  const isLoading = useLoading()
-  const { setPersistentFocus } = useFocusContext()
-  const { browseType, setBrowseType } = useBrowseContext()
-
-  const [backUrl, setBackUrl] = useState<string | null>(null)
-  const [searchTerm, setSearchTerm] = useState((router.query.q as string) || "")
-
-  useEffect(() => {
-    if (typeof window !== "undefined") {
-      const ref = document.referrer
-      if (ref.includes("/browse?q")) {
-        setBackUrl(ref)
-      }
-    }
-  }, [])
-
-  const scopeFromQuery = (router.query.search_scope as string) || "has"
-  const selectedOption = getBrowseFormKey(browseType, scopeFromQuery)
-
-  const optionData = selectOptions[selectedOption]
-  const searchScope = optionData.scope
-  const placeholder = optionData.placeholder
-  const tipText = optionData.searchTip
-
-  const formattedSelectOptions = Object.keys(selectOptions).map((key) => ({
-    text: selectOptions[key].text,
-    value: key,
-  }))
-
-  const handleChange = (e: SyntheticEvent, setValue: (val: string) => void) => {
-    const target = e.target as HTMLInputElement
-    setValue(target.value)
-  }
-
-  const handleSubmit = async (e: SyntheticEvent) => {
-    e.preventDefault()
-
-    const basePath = browseType === "subjects" ? "/browse" : "/browse/authors"
-    setBrowseType(browseType)
-
-    const queryString = getQueryString({
-      q: searchTerm,
-      search_scope: searchScope,
-    })
-
-    setPersistentFocus(onSubmitFocusId)
-    await router.push(`${basePath}${queryString}`)
-  }
-
-  return (
-    <div className={`${styles.searchContainer} no-print`}>
-      <Box
-        sx={{
-          margin: "0 auto",
-          maxWidth: "1280px",
-          px: { base: "s", md: "m", xl: "s" },
-        }}
-      >
-        <Text size="body2" className={styles.searchTip}>
-          <Icon size="medium" name="errorOutline" iconRotation="rotate180" />
-          <Box as="span" className={styles.searchTipText}>
-            <span className={styles.searchTipTitle}>{tipTitle}</span>
-            {tipText}
-          </Box>
-        </Text>
-
-        <SearchBar
-          id="mainContent"
-          action={path}
-          method="get"
-          onSubmit={handleSubmit}
-          labelText={labelText}
-          isDisabled={isLoading}
-          pb={{ base: children ? 0 : "l", md: 0 }}
-          selectProps={{
-            value: selectedOption,
-            labelText: "Select a category",
-            name: "browse_option",
-            optionsData: formattedSelectOptions,
-            onChange: (e) => {
-              const newValue = (e.target as HTMLSelectElement).value
-              const newOption = selectOptions[newValue]
-
-              setBrowseType(newOption.browseType)
-
-              const newBasePath =
-                newOption.browseType === "subjects"
-                  ? "/browse"
-                  : "/browse/authors"
-
-              router.push(
-                {
-                  pathname: newBasePath,
-                  query: { search_scope: newOption.scope },
-                },
-                undefined,
-                { shallow: true, scroll: false }
-              )
-            },
-          }}
-          textInputProps={{
-            isClearable: true,
-            onChange: (e) => handleChange(e, setSearchTerm),
-            isClearableCallback: () => setSearchTerm(""),
-            value: searchTerm,
-            name: "q",
-            placeholder,
-            labelText: tipText,
-          }}
-        />
-
-        {(children ||
-          activePage === "sh-results" ||
-          activePage === "contributor-results") && (
-          <Flex
-            direction="column"
-            justifyContent="space-between"
-            mt={{ base: "0", md: "xs" }}
-          >
-            {activePage === "sh-results" && backUrl && (
-              <Link
-                variant="buttonSecondary"
-                id="back-index"
-                width={{ base: "100%", md: "fit-content" }}
-                onClick={() => router.push(backUrl)}
-                gap="xxs"
-                mt="xxs"
-                background="white"
-                mb={{ base: "xs", md: 0 }}
-              >
-                Back to index
-              </Link>
-            )}
-            {children}
-          </Flex>
-        )}
-      </Box>
-    </div>
-  )
-}
-
-export default SearchBrowseForm
