+++ src/components/SearchForm/SearchForm.tsx
@@ -22,17 +22,21 @@ import { idConstants, useFocusContext } from "../../context/FocusContext"
 import type { Aggregation } from "../../types/filterTypes"
 import { collapseMultiValueQueryParams } from "../../utils/refineSearchUtils"
 import { getSearchQuery } from "../../utils/searchUtils"
+import type { RCPage } from "../../types/pageTypes"
 
 const SearchForm = ({
+  activePage,
   aggregations,
   searchResultsCount,
 }: {
+  activePage: RCPage
   aggregations?: Aggregation[]
   searchResultsCount?: number
 }) => {
   const router = useRouter()
   const isLoading = useLoading()
   const { setPersistentFocus } = useFocusContext()
+  const [backUrl, setBackUrl] = useState<string | null>(null)
   const [searchTerm, setSearchTerm] = useState((router.query.q as string) || "")
   const [searchScope, setSearchScope] = useState(
     (router?.query?.search_scope as string) || "all"
@@ -45,6 +49,11 @@ const SearchForm = ({
     setAppliedFilters(collapseMultiValueQueryParams(router.query))
     setSearchScope((router.query.search_scope as string) || "all")
     setSearchTerm((router.query.q as string) || "")
+
+    if (typeof window !== "undefined") {
+      const ref = document.referrer
+      if (ref.includes("/browse?q")) setBackUrl(ref)
+    }
   }, [router.query])
 
   const displayFilters = !!aggregations?.filter((agg) => agg.values.length)
@@ -102,9 +111,8 @@ const SearchForm = ({
           action={PATHS.SEARCH}
           method="get"
           onSubmit={handleSubmit}
-          labelText="Search Bar"
+          labelText="Search Bar Label"
           isDisabled={isLoading}
-          pb="xs"
           selectProps={{
             value: searchScope,
             labelText: "Select a category",
@@ -122,7 +130,39 @@ const SearchForm = ({
             labelText: tipText,
           }}
         />
-        <Flex direction="column">
+        <Flex
+          direction="column"
+          justifyContent="space-between"
+          mt={{ base: "0", md: "xs" }}
+        >
+          {(activePage === "sh-results" ||
+            activePage === "contributor-results") &&
+            backUrl && (
+              <Link
+                variant="buttonSecondary"
+                id="back-index"
+                width={{ base: "100%", md: "fit-content" }}
+                onClick={() => router.push(backUrl)}
+                gap="xxs"
+                mt="xxs"
+                background="white"
+                mb={{ base: "xs", md: 0 }}
+              >
+                <svg
+                  width="24"
+                  height="24"
+                  viewBox="0 0 24 24"
+                  fill="none"
+                  xmlns="http://www.w3.org/2000/svg"
+                >
+                  <path
+                    d="M20 11H7.83L13.42 5.41L12 4L4 12L12 20L13.41 18.59L7.83 13H20V11Z"
+                    fill="#0069BF"
+                  />
+                </svg>
+                Back to index
+              </Link>
+            )}
           <Link
             className={styles.advancedSearch}
             href="/search/advanced"
