+++ src/utils/testUtils.tsx
@@ -2,11 +2,14 @@ import React, { type ReactElement } from "react"
 import { render, type RenderOptions } from "@testing-library/react"
 import { FeedbackProvider } from "../context/FeedbackContext"
 import { FocusProvider } from "../context/FocusContext"
+import { BrowseProvider } from "../context/BrowseContext"
 
 const AllTheProviders = ({ children }: { children: React.ReactNode }) => {
   return (
     <FeedbackProvider value={null}>
-      <FocusProvider>{children}</FocusProvider>
+      <FocusProvider>
+        <BrowseProvider>{children}</BrowseProvider>
+      </FocusProvider>
     </FeedbackProvider>
   )
 }
